information hiding of split functionality in language_helper.py
